(PC-4517) sandboxes: Move utility functions outside of tests
Two scripts in the "sandboxes" package uses functions that are defined
in tests. This will not work once we remove the PYTHONPATH override
and tests are not installed anymore.
